+++ Jellyfin.Data/Entities/ActivityLog.cs
@@ -5,34 +5,18 @@ using Microsoft.Extensions.Logging;
 
 namespace Jellyfin.Data.Entities
 {
-    public partial class ActivityLog
+    /// <summary>
+    /// An entity referencing an activity log entry.
+    /// </summary>
+    public partial class ActivityLog : ISavingChanges
     {
-        partial void Init();
-
         /// <summary>
-        /// Default constructor. Protected due to required properties, but present because EF needs it.
+        /// Initializes a new instance of the <see cref="ActivityLog"/> class.
+        /// Public constructor with required data.
         /// </summary>
-        protected ActivityLog()
-        {
-            Init();
-        }
-
-        /// <summary>
-        /// Replaces default constructor, since it's protected. Caller assumes responsibility for setting all required values before saving.
-        /// </summary>
-        public static ActivityLog CreateActivityLogUnsafe()
-        {
-            return new ActivityLog();
-        }
-
-        /// <summary>
-        /// Public constructor with required data
-        /// </summary>
-        /// <param name="name"></param>
-        /// <param name="type"></param>
-        /// <param name="userId"></param>
-        /// <param name="datecreated"></param>
-        /// <param name="logSeverity"></param>
+        /// <param name="name">The name.</param>
+        /// <param name="type">The type.</param>
+        /// <param name="userId">The user id.</param>
         public ActivityLog(string name, string type, Guid userId)
         {
             if (string.IsNullOrEmpty(name))
@@ -55,13 +39,20 @@ namespace Jellyfin.Data.Entities
         }
 
         /// <summary>
+        /// Initializes a new instance of the <see cref="ActivityLog"/> class.
+        /// Default constructor. Protected due to required properties, but present because EF needs it.
+        /// </summary>
+        protected ActivityLog()
+        {
+            Init();
+        }
+
+        /// <summary>
         /// Static create function (for use in LINQ queries, etc.)
         /// </summary>
-        /// <param name="name"></param>
-        /// <param name="type"></param>
-        /// <param name="userId"></param>
-        /// <param name="datecreated"></param>
-        /// <param name="logseverity"></param>
+        /// <param name="name">The name.</param>
+        /// <param name="type">The type.</param>
+        /// <param name="userId">The user's id.</param>
         public static ActivityLog Create(string name, string type, Guid userId)
         {
             return new ActivityLog(name, type, userId);
@@ -72,7 +63,8 @@ namespace Jellyfin.Data.Entities
          *************************************************************************/
 
         /// <summary>
-        /// Identity, Indexed, Required
+        /// Gets the identity of this instance.
+        /// This is the key in the backing database.
         /// </summary>
         [Key]
         [Required]
@@ -80,7 +72,8 @@ namespace Jellyfin.Data.Entities
         public int Id { get; protected set; }
 
         /// <summary>
-        /// Required, Max length = 512
+        /// Gets or sets the name.
+        /// Required, Max length = 512.
         /// </summary>
         [Required]
         [MaxLength(512)]
@@ -88,21 +81,24 @@ namespace Jellyfin.Data.Entities
         public string Name { get; set; }
 
         /// <summary>
-        /// Max length = 512
+        /// Gets or sets the overview.
+        /// Max length = 512.
         /// </summary>
         [MaxLength(512)]
         [StringLength(512)]
         public string Overview { get; set; }
 
         /// <summary>
-        /// Max length = 512
+        /// Gets or sets the short overview.
+        /// Max length = 512.
         /// </summary>
         [MaxLength(512)]
         [StringLength(512)]
         public string ShortOverview { get; set; }
 
         /// <summary>
-        /// Required, Max length = 256
+        /// Gets or sets the type.
+        /// Required, Max length = 256.
         /// </summary>
         [Required]
         [MaxLength(256)]
@@ -110,41 +106,48 @@ namespace Jellyfin.Data.Entities
         public string Type { get; set; }
 
         /// <summary>
-        /// Required
+        /// Gets or sets the user id.
+        /// Required.
         /// </summary>
         [Required]
         public Guid UserId { get; set; }
 
         /// <summary>
-        /// Max length = 256
+        /// Gets or sets the item id.
+        /// Max length = 256.
         /// </summary>
         [MaxLength(256)]
         [StringLength(256)]
         public string ItemId { get; set; }
 
         /// <summary>
-        /// Required
+        /// Gets or sets the date created. This should be in UTC.
+        /// Required.
         /// </summary>
         [Required]
         public DateTime DateCreated { get; set; }
 
         /// <summary>
-        /// Required
+        /// Gets or sets the log severity. Default is <see cref="LogLevel.Trace"/>.
+        /// Required.
         /// </summary>
         [Required]
         public LogLevel LogSeverity { get; set; }
 
         /// <summary>
+        /// Gets or sets the row version.
         /// Required, ConcurrencyToken.
         /// </summary>
         [ConcurrencyCheck]
         [Required]
         public uint RowVersion { get; set; }
 
+        partial void Init();
+
+        /// <inheritdoc />
         public void OnSavingChanges()
         {
             RowVersion++;
         }
     }
 }
-
