+++ Jellyfin.Server.Implementations/JellyfinDb.cs
@@ -110,7 +110,7 @@ namespace Jellyfin.Server.Implementations
             foreach (var entity in ChangeTracker.Entries().Where(e => e.State == EntityState.Modified))
             {
                 var saveEntity = entity.Entity as ISavingChanges;
-                saveEntity.OnSavingChanges();
+                saveEntity?.OnSavingChanges();
             }
 
             return base.SaveChanges();
