+++ MediaBrowser.Api/System/ActivityLogService.cs
@@ -1,3 +1,7 @@
+using System;
+using System.Globalization;
+using System.Linq;
+using Jellyfin.Data.Entities;
 using MediaBrowser.Controller.Configuration;
 using MediaBrowser.Controller.Net;
 using MediaBrowser.Model.Activity;
@@ -47,7 +51,14 @@ namespace MediaBrowser.Api.System
 
         public object Get(GetActivityLogs request)
         {
-            var result = _activityManager.GetPagedResult(request.StartIndex, request.Limit);
+            DateTime? minDate = string.IsNullOrWhiteSpace(request.MinDate) ?
+                (DateTime?)null :
+                DateTime.Parse(request.MinDate, null, DateTimeStyles.RoundtripKind).ToUniversalTime();
+
+            var filterFunc = new Func<IQueryable<ActivityLog>, IQueryable<ActivityLog>>(
+                entries => entries.Where(entry => entry.DateCreated >= minDate));
+
+            var result = _activityManager.GetPagedResult(filterFunc, request.StartIndex, request.Limit);
 
             return ToOptimizedResult(result);
         }
