+++ MediaBrowser.Model/Activity/IActivityManager.cs
@@ -1,7 +1,6 @@
 #pragma warning disable CS1591
 
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Jellyfin.Data.Entities;
@@ -21,7 +20,7 @@ namespace MediaBrowser.Model.Activity
         QueryResult<ActivityLogEntry> GetPagedResult(int? startIndex, int? limit);
 
         QueryResult<ActivityLogEntry> GetPagedResult(
-            Func<IQueryable<ActivityLog>, IEnumerable<ActivityLog>> func,
+            Func<IQueryable<ActivityLog>, IQueryable<ActivityLog>> func,
             int? startIndex,
             int? limit);
     }
