+++ MediaBrowser.Model/Devices/DeviceOptions.cs
@@ -1,0 +1,9 @@
+#pragma warning disable CS1591
+
+namespace MediaBrowser.Model.Devices
+{
+    public class DeviceOptions
+    {
+        public string CustomName { get; set; }
+    }
+}
